added logos, got all info needed for main page, added footer
