+++ client/store/index.js
@@ -5,8 +5,14 @@ import {composeWithDevTools} from 'redux-devtools-extension'
 import singleHedgeFund from './oneFund'
 import hedgeFunds from './funds'
 import stockStats from './stockStats'
+import hedgeStats from './hedgeStats'
 
-const reducer = combineReducers({singleHedgeFund, hedgeFunds, stockStats})
+const reducer = combineReducers({
+  singleHedgeFund,
+  hedgeFunds,
+  stockStats,
+  hedgeStats,
+})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
