+++ public/style.css
@@ -111,7 +111,8 @@ form div {
 
 .explainer {
   text-align: center;
-  font-style: italic;
+  /* font-style: italic; */
+  font-weight: 700;
 }
 
 /* ALL HEDGE FUNDS STYLING */
@@ -333,4 +334,11 @@ tr:nth-child(even) {
 .pie-chart {
   border-radius: 15px;
   box-shadow: 1px 2px 5px rgba(0, 0, 0, 0.65);
+}
+
+/* Navbar */
+
+#Nav-Logo {
+  height: 100%;
+  width: 250px;
 }
