+++ server/api/index.js
@@ -3,6 +3,7 @@ module.exports = router
 
 router.use('/hedgefunds', require('./hedgefunds'))
 router.use('/stockStats', require('./stockStats'))
+router.use('/hedgeStats', require('./hedgeStats'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
